added task card

+++ src/main/java/by/itacademy/todolist/controller/command/UpdateTaskCommand.java
@@ -34,7 +34,9 @@ public class UpdateTaskCommand extends FrontCommand {
                     task.setCompleted(true);
                     break;
                 case ApplicationConstants.TASK_ACTION_FULL_DELETE:
-                    fileService.delete(task.getFileInfo().getId());
+                    if (task.getFileInfo().getPath() != null) {
+                        fileService.delete(task.getFileInfo().getId());
+                    }
                     taskService.deleteTask(takId);
                     sectionTasksViewCommand.process();
                     return;
